breaking eplant layout into its own file
